+++ app/Http/Controllers/SubProductController.php
@@ -15,7 +15,7 @@ class SubProductController extends Controller
      */
     public function index(Request $request)
     {
-
+        $this->authorize('v_sub_product');
         $subProducts = SubProduct::all();
 
         return view('pages.subProduct.index', compact('subProducts'));
@@ -27,6 +27,7 @@ class SubProductController extends Controller
      */
     public function create(Request $request)
     {
+        $this->authorize('c_sub_product');
         return view('pages.subProduct.create');
     }
 
@@ -36,6 +37,7 @@ class SubProductController extends Controller
      */
     public function store(SubProductStoreRequest $request)
     {
+        $this->authorize('c_sub_product');
         $subProduct = SubProduct::create($request->validated());
 
         $request->session()->flash('subProduct.create', $subProduct->name .' create SuccessFully');
@@ -50,6 +52,7 @@ class SubProductController extends Controller
      */
     public function show(Request $request, SubProduct $subProduct)
     {
+        $this->authorize('v_sub_product');
         return view('pages.subProduct.show', compact('subProduct'));
     }
 
@@ -60,6 +63,7 @@ class SubProductController extends Controller
      */
     public function edit(SubProduct $subProduct)
     {
+        $this->authorize('u_sub_product');
         return view('pages.subProduct.edit', compact('subProduct'));
     }
 
@@ -70,6 +74,7 @@ class SubProductController extends Controller
      */
     public function update(SubProductUpdateRequest $request, SubProduct $subProduct)
     {
+        $this->authorize('u_sub_product');
         $valid = array_filter($request->validated());
         $subProduct->update($valid);
 
@@ -85,6 +90,7 @@ class SubProductController extends Controller
      */
     public function destroy(Request $request, SubProduct $subProduct)
     {
+        $this->authorize('d_sub_product');
         $subProduct->delete();
 
         return redirect()->route('sub-product.index');
